specify hori_buffer_len dtype

+++ horayzon/horizon.pyx
@@ -201,7 +201,7 @@ def horizon_gridded(
         hori_buffer_len = vec_norm.shape[0] * vec_norm.shape[1] * azim_num
     else:
         print("Horizon buffer size is restricted")
-        hori_buffer_len = int((hori_buffer_size_max * 10 ** 9) / 4) + 100000
+        hori_buffer_len = np.int64((hori_buffer_size_max * 10 ** 9) / 4) + 100000
         # add some "safety" memory to the buffer (-> 100000)
     cdef np.ndarray[np.float32_t, ndim = 1, mode = "c"] \
         hori_buffer = np.empty(hori_buffer_len,  dtype=np.float32)
